dmp-125 : Restructure of Java Package
