+++ src/integrationTest/java/uk/gov/hmcts/darts/common/controller/GetWelcomeTest.java
@@ -1,0 +1,27 @@
+package uk.gov.hmcts.darts.common.controller;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.MvcResult;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@WebMvcTest
+class GetWelcomeTest {
+
+    @Autowired
+    private transient MockMvc mockMvc;
+
+    @DisplayName("Should welcome upon root request with 200 response code")
+    @Test
+    void welcomeRootEndpoint() throws Exception {
+        MvcResult response = mockMvc.perform(get("/")).andExpect(status().isOk()).andReturn();
+
+        assertThat(response.getResponse().getContentAsString()).startsWith("Welcome");
+    }
+}
