+++ src/main/java/uk/gov/hmcts/darts/Application.java
@@ -1,0 +1,17 @@
+package uk.gov.hmcts.darts;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
+import org.springframework.boot.autoconfigure.security.servlet.UserDetailsServiceAutoConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+
+@SpringBootApplication(exclude = {SecurityAutoConfiguration.class, UserDetailsServiceAutoConfiguration.class})
+@ComponentScan({"uk.gov.hmcts.darts"})
+@SuppressWarnings("HideUtilityClassConstructor") // Spring needs a constructor, its not a utility class
+public class Application {
+
+    public static void main(final String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+}
