+++ src/main/java/uk/gov/hmcts/darts/authentication/service/AuthenticationHelperService.java
@@ -1,0 +1,10 @@
+package uk.gov.hmcts.darts.authentication.service;
+
+import groovy.util.logging.Slf4j;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@Component
+public class AuthenticationHelperService {
+
+}
