+++ src/main/java/uk/gov/hmcts/darts/authentication/service/AuthenticationService.java
@@ -1,0 +1,7 @@
+package uk.gov.hmcts.darts.authentication.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class AuthenticationService {
+}
