+++ src/main/java/uk/gov/hmcts/darts/authentication/service/SessionService.java
@@ -1,0 +1,12 @@
+package uk.gov.hmcts.darts.authentication.service;
+
+import groovy.util.logging.Slf4j;
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@Component
+@AllArgsConstructor
+public class SessionService {
+
+}
