+++ src/main/java/uk/gov/hmcts/darts/authentication/util/AuthenticationType.java
@@ -1,0 +1,6 @@
+package uk.gov.hmcts.darts.authentication.util;
+
+public enum AuthenticationType {
+    AD,
+    B2C
+}
