--- src/main/java/uk/gov/hmcts/reform/darts/controllers/authentication/AuthExternalUserController.java
@@ -1,19 +1,0 @@
-package uk.gov.hmcts.reform.darts.controllers.authentication;
-
-import groovy.util.logging.Slf4j;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.ModelAndView;
-
-@Slf4j
-@RestController
-@RequestMapping("/external-user")
-public class AuthExternalUserController extends AuthController {
-
-    @GetMapping("/login-or-refresh")
-    public ModelAndView loginOrRefresh() {
-
-        return new ModelAndView();
-    }
-}
