--- src/main/java/uk/gov/hmcts/reform/darts/services/authentication/AuthHelperService.java
@@ -1,10 +1,0 @@
-package uk.gov.hmcts.reform.darts.services.authentication;
-
-import groovy.util.logging.Slf4j;
-import org.springframework.stereotype.Component;
-
-@Slf4j
-@Component
-public class AuthHelperService {
-
-}
