--- src/main/java/uk/gov/hmcts/reform/darts/services/authentication/AuthService.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.reform.darts.services.authentication;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class AuthService {
-}
