--- src/main/java/uk/gov/hmcts/reform/darts/services/authentication/SessionService.java
@@ -1,12 +1,0 @@
-package uk.gov.hmcts.reform.darts.services.authentication;
-
-import groovy.util.logging.Slf4j;
-import lombok.AllArgsConstructor;
-import org.springframework.stereotype.Component;
-
-@Slf4j
-@Component
-@AllArgsConstructor
-public class SessionService {
-
-}
