--- src/main/java/uk/gov/hmcts/reform/darts/utils/AuthType.java
@@ -1,6 +1,0 @@
-package uk.gov.hmcts.reform.darts.utils;
-
-public enum AuthType {
-    AD,
-    B2C
-}
