+++ src/smokeTest/java/uk/gov/hmcts/darts/SampleSmokeTest.java
@@ -1,0 +1,39 @@
+package uk.gov.hmcts.darts;
+
+import io.restassured.RestAssured;
+import io.restassured.http.ContentType;
+import io.restassured.response.Response;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import static io.restassured.RestAssured.given;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+class SampleSmokeTest {
+    protected static final String CONTENT_TYPE_VALUE = "application/json";
+
+    @Value("${TEST_URL:http://localhost:8080}")
+    private String testUrl;
+
+    @BeforeEach
+    public void setUp() {
+        RestAssured.baseURI = testUrl;
+        RestAssured.useRelaxedHTTPSValidation();
+    }
+
+    @Test
+    void smokeTest() {
+        Response response = given()
+            .contentType(ContentType.JSON)
+            .when()
+            .get()
+            .then()
+            .extract().response();
+
+        Assertions.assertEquals(200, response.statusCode());
+        Assertions.assertTrue(response.asString().startsWith("Welcome"));
+    }
+}
